Move schemas to schema folder
