+++ packages/apollos-data-connector-rock-content/src/content-items/index.js
@@ -24,111 +24,7 @@ const createImageUrl = (uri) =>
 
 // export { default as model } from './model';
 export { default as dataSource } from './data-source';
-
-export const schema = gql`
-  type SharableContentItem implements Sharable {
-    url: String
-    message: String
-    title: String
-  }
-
-  interface ContentItem {
-    id: ID!
-    title: String
-    coverImage: ImageMedia
-    images: [ImageMedia]
-    videos: [VideoMedia]
-    audios: [AudioMedia]
-    htmlContent: String
-    childContentItemsConnection(
-      first: Int
-      after: String
-    ): ContentItemsConnection
-    siblingContentItemsConnection(
-      first: Int
-      after: String
-    ): ContentItemsConnection
-    parentChannel: ContentChannel
-
-    sharing: SharableContentItem
-    theme: Theme
-    isLiked: Boolean
-  }
-
-  type UniversalContentItem implements ContentItem & Node {
-    id: ID!
-    title: String
-    coverImage: ImageMedia
-    images: [ImageMedia]
-    videos: [VideoMedia]
-    audios: [AudioMedia]
-    htmlContent: String
-    childContentItemsConnection(
-      first: Int
-      after: String
-    ): ContentItemsConnection
-    siblingContentItemsConnection(
-      first: Int
-      after: String
-    ): ContentItemsConnection
-    parentChannel: ContentChannel
-    terms(match: String): [Term]
-
-    sharing: SharableContentItem
-    theme: Theme
-    isLiked: Boolean
-  }
-
-  type DevotionalContentItem implements ContentItem & Node {
-    id: ID!
-    title: String
-    coverImage: ImageMedia
-    images: [ImageMedia]
-    videos: [VideoMedia]
-    audios: [AudioMedia]
-    htmlContent: String
-    childContentItemsConnection(
-      first: Int
-      after: String
-    ): ContentItemsConnection
-    siblingContentItemsConnection(
-      first: Int
-      after: String
-    ): ContentItemsConnection
-    parentChannel: ContentChannel
-
-    sharing: SharableContentItem
-    theme: Theme
-    isLiked: Boolean
-    scriptures: [Scripture]
-  }
-
-  type Term {
-    key: String
-    value: String
-  }
-
-  input ContentItemsConnectionInput {
-    first: Int
-    after: String
-  }
-
-  type ContentItemsConnection {
-    edges: [ContentItemsConnectionEdge]
-    # TODO totalCount: Int
-    pageInfo: PaginationInfo
-  }
-
-  type ContentItemsConnectionEdge {
-    node: ContentItem
-    cursor: String
-  }
-
-  extend type Query {
-    userFeed(first: Int, after: String): ContentItemsConnection
-    getAllLikedContent: [ContentItem]
-  }
-`;
+export { contentItemSchema as schema } from '@apollosproject/data-schema';
 
 // Empty fields in rock default to `''`
 const hasScripture = ({ attributeValues }) =>
